+++ packages/authorize/index.d.ts
@@ -1,2 +1,3 @@
 export { default } from './Authorize';
 export * from './Authorize';
+export * from './useAuthorize';