+++ packages/reactstrap-validation-date/AvDate.d.ts
@@ -1,6 +1,7 @@
 import { AvInput } from "./AvInput";
 
 export interface AvDateProps extends AvInput{
+    hideIcon?: boolean;
 }
 
 declare const AvDate: React.ComponentType<AvDateProps>;