+++ packages/spaces/src/Spaces.js
@@ -40,6 +40,24 @@ export const getAllSpaces = async (
   }
 };
 
+export const sanitizeSpaces = spaces => {
+  // Normalize space pairs ( [{ name, value }} => { name: value } )
+  const pairFields = ['images', 'metadata', 'colors', 'icons', 'mapping'];
+  return spaces.reduce((accum, spc) => {
+    pairFields.forEach(field => {
+      if (spc[field] && Array.isArray(spc[field])) {
+        spc[field] = spc[field].reduce((_accum, { name, value }) => {
+          _accum[name] = value;
+          return _accum;
+        }, {});
+      }
+    });
+
+    accum.push(spc);
+    return accum;
+  }, []);
+};
+
 export const SpacesContext = createContext();
 
 const Spaces = ({
@@ -49,22 +67,32 @@ const Spaces = ({
   spaceIds,
   payerIds,
   children,
+  spaces: spacesFromProps,
 }) => {
   const [spaces, setSpaces] = useState([]);
+  const [loading, setLoading] = useState(true);
 
   // NOTE: we do not want to query slotmachine by payerIDs and spaceIDs at the same time
   // because slotmachine does an AND on those conditions. We want OR
   useEffectAsync(async () => {
+    setLoading(true);
     // Filter out dupes and ids that we already have the space for
     const filteredSpaceIDs = spaceIds
       .filter((id, i) => spaceIds.indexOf(id) === i)
-      .filter(id => !spaces.some(spc => spc && spc.id === id));
+      .filter(id => !spaces.some(spc => spc && spc.id === id))
+      .filter(id => !spacesFromProps.some(spc => spc && spc.id === id));
 
     const filteredPayerIDs = payerIds
       .filter((id, i) => payerIds.indexOf(id) === i)
       .filter(
         id =>
           !spaces.some(
+            spc => spc && spc.payerIDs && spc.payerIDs.some(pId => pId === id)
+          )
+      )
+      .filter(
+        id =>
+          !spacesFromProps.some(
             spc => spc && spc.payerIDs && spc.payerIDs.some(pId => pId === id)
           )
       );
@@ -92,38 +120,37 @@ const Spaces = ({
       _spaces = _spaces.concat(spacesByPayerIDs);
     }
 
-    // Normalize space pairs ( [{ name, value }} => { name: value } )
-    const pairFields = ['images', 'metadata', 'colors', 'icons', 'mapping'];
-    _spaces = _spaces.reduce((accum, spc) => {
-      pairFields.forEach(field => {
-        if (spc[field]) {
-          spc[field] = spc[field].reduce((_accum, { name, value }) => {
-            _accum[name] = value;
-            return _accum;
-          }, {});
-        }
-      });
-
-      accum.push(spc);
-      return accum;
-    }, []);
-
     if (_spaces.length > 0) setSpaces(_spaces);
+    setLoading(false);
   }, [payerIds, spaceIds]);
 
+  const spacesForProvider = sanitizeSpaces(spaces.concat(spacesFromProps));
   return (
-    <SpacesContext.Provider value={{ spaces }}>
+    <SpacesContext.Provider value={{ spaces: spacesForProvider, loading }}>
       {children}
     </SpacesContext.Provider>
   );
 };
 
 export const useSpace = id => {
-  const { spaces = [] } = useContext(SpacesContext) || {};
+  const { spaces = [], loading } = useContext(SpacesContext) || {};
 
   // Try to match by space id first, else match by payer id
   // eslint-disable-next-line react-hooks/exhaustive-deps
   const space = useMemo(() => {
+    // If we don't pass a spaceId in then we will get the first space in the array. If there is more than one space we will raise a
+    // warning because it should only be expected that we use no spaceId if the app only is using a single space in the provider.
+    if (id === undefined) {
+      if (spaces.length > 1) {
+        // eslint-disable-next-line no-console
+        console.warn(
+          `You did not pass an ID in to find a space, and there is more than 1 space in the space array. Returning the first.`
+        );
+      }
+
+      return spaces[0];
+    }
+
     let [spc] = spaces.filter(s => s.id === id);
 
     if (!spc) {
@@ -150,7 +177,7 @@ export const useSpace = id => {
     );
   });
 
-  return { space, isGhost };
+  return { space, isGhost, loading };
 };
 
 Spaces.propTypes = {
@@ -160,6 +187,7 @@ Spaces.propTypes = {
   variables: PropTypes.object,
   spaceIds: PropTypes.arrayOf(PropTypes.string),
   payerIds: PropTypes.arrayOf(PropTypes.string),
+  spaces: PropTypes.arrayOf(PropTypes.object),
 };
 
 Spaces.defaultProps = {
@@ -190,6 +218,7 @@ Spaces.defaultProps = {
   variables: { types: ['space'] },
   spaceIds: [],
   payerIds: [],
+  spaces: [],
 };
 
 export default Spaces;