+++ packages/spaces/CHANGELOG.md
@@ -3,6 +3,71 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.0.3](https://github.com/Availity/availity-react/compare/@availity/spaces@2.0.2...@availity/spaces@2.0.3) (2019-07-11)
+
+
+### Bug Fixes
+
+* **spaces:** allow number or string for skeleton width/height ([800c937](https://github.com/Availity/availity-react/commit/800c937))
+* **spaces:** only spread skeletonProps onto Loader ([d2aa0af](https://github.com/Availity/availity-react/commit/d2aa0af))
+
+
+
+
+
+## [2.0.2](https://github.com/Availity/availity-react/compare/@availity/spaces@2.0.1...@availity/spaces@2.0.2) (2019-07-02)
+
+**Note:** Version bump only for package @availity/spaces
+
+
+
+
+
+## [2.0.1](https://github.com/Availity/availity-react/compare/@availity/spaces@2.0.0...@availity/spaces@2.0.1) (2019-06-24)
+
+**Note:** Version bump only for package @availity/spaces
+
+
+
+
+
+# [2.0.0](https://github.com/Availity/availity-react/compare/@availity/spaces@1.3.0...@availity/spaces@2.0.0) (2019-06-20)
+
+
+### Bug Fixes
+
+* **spaces:** wait for spaces to load to default url in SpacesLogo ([02bf3e4](https://github.com/Availity/availity-react/commit/02bf3e4))
+
+
+### Code Refactoring
+
+* **spaces:** making breaking change commit to major bump spaces ([4134486](https://github.com/Availity/availity-react/commit/4134486))
+
+
+### Features
+
+* **spaces:** render spinner til space loads in SpacesImage ([e9acb79](https://github.com/Availity/availity-react/commit/e9acb79))
+
+
+### BREAKING CHANGES
+
+* **spaces:** [Commit](https://github.com/Availity/availity-react/pull/170/commits/e9acb794c504ec159949a7fe82ea76c9e4f5c986) was noted as a breaking change but not using the correct conventional commit format for lerna to major bump the package. reactstrap >= 8 is now a peer dep
+
+
+
+
+
+# [1.3.0](https://github.com/Availity/availity-react/compare/@availity/spaces@1.2.0...@availity/spaces@1.3.0) (2019-06-14)
+
+
+### Features
+
+* **spaces:** add spaces prop ([ce7d721](https://github.com/Availity/availity-react/commit/ce7d721))
+
+
+
+
+
 # [1.2.0](https://github.com/Availity/availity-react/compare/@availity/spaces@1.1.1...@availity/spaces@1.2.0) (2019-06-06)